feat(core): adding tray to instances list
Created tray instances will now show in the ps list.
