+++ packages/desktop-core/src/main/instances/index.ts
@@ -1,10 +1,16 @@
 import { CompositeInstanceService } from "./compositeInstanceService";
 import { WindowInstanceService } from "./windowInstanceService";
 import { StorageInstanceService } from "./storageInstanceService";
+import { TrayInstanceService } from "./trayInstanceService";
 
 export * from "./iInstanceService";
 
 const windowInstanceService = new WindowInstanceService();
 const storageInstanceService = new StorageInstanceService();
+const trayInstanceService = new TrayInstanceService();
 
-export const instanceService = new CompositeInstanceService(windowInstanceService, storageInstanceService);
+export const instanceService = new CompositeInstanceService(
+    windowInstanceService,
+    storageInstanceService,
+    trayInstanceService,
+);
