+++ packages/desktop-core/src/main/tray/factory/iTrayFactory.ts
@@ -1,0 +1,6 @@
+import { IConfiguration } from "@reactivemarkets/desktop-types";
+import { Tray } from "electron";
+
+export interface ITrayFactory {
+    create(configuration: IConfiguration): Promise<Tray>;
+}
