+++ packages/desktop-core/src/main/tray/factory/index.ts
@@ -1,0 +1,2 @@
+export * from "./defaultTrayFactory";
+export * from "./iTrayFactory";
