+++ packages/desktop-types/src/configuration/iConfiguration.ts
@@ -1,15 +1,16 @@
 import { ConfigurationKind } from "./configurationKind";
 import { IApplicationSpecification } from "./iApplicationSpecification";
+import { IApplicationStatus } from "./iApplicationStatus";
 import { IExternalSpecification } from "./iExternalSpecification";
+import { IExternalStatus } from "./iExternalStatus";
 import { IMetadata } from "./iMetadata";
 import { IServiceSpecification } from "./iServiceSpecification";
+import { IServiceStatus } from "./iServiceStatus";
 import { ISessionSpecification } from "./iSessionSpecification";
 import { IStorageSpecification } from "./iStorageSpecification";
-import { ITraySpecification } from "./iTraySpecification";
-import { IApplicationStatus } from "./iApplicationStatus";
-import { IExternalStatus } from "./iExternalStatus";
-import { IServiceStatus } from "./iServiceStatus";
 import { IStorageStatus } from "./iStorageStatus";
+import { ITraySpecification } from "./iTraySpecification";
+import { ITrayStatus } from "./iTrayStatus";
 
 type Spec =
     | IApplicationSpecification
@@ -19,7 +20,7 @@ type Spec =
     | IStorageSpecification
     | ITraySpecification;
 
-type Status = IApplicationStatus | IExternalStatus | IServiceStatus | IStorageStatus;
+type Status = IApplicationStatus | IExternalStatus | IServiceStatus | IStorageStatus | ITrayStatus;
 
 export interface IConfiguration {
     readonly apiVersion?: string;
