+++ packages/desktop-types/src/configuration/iTrayStatus.ts
@@ -1,0 +1,3 @@
+export interface ITrayStatus {
+    readonly startTime?: Date;
+}
